refactor(pdf): extract publication metadata
